6 (chore): Add dependencies for tests, webpack configs and async await
Add enzyme to take advantage of it shallow component build. Add also
enzyme-to-json for better snapshot comparison.
Add webpack-merge in order to be able to differentiate between webpack
configuration for development and for production. This allow to split
the configuration in 2 files, update the npm scripts accordingly.
Add babel-polyfil due to an error occurring when defining a class method
as async.